pyproject.toml
[project]
name = "astradesk"
version = "0.2.1"
description = "AstraDesk – internal AI agents framework (Python core)"
requires-python = ">=3.11,<3.12"
authors = [{name = "Siergej Sobolewski", email = "s.sobolewski@hotmail.com"}]
license = {text = "Apache-2.0"}

dependencies = [
  "fastapi>=0.115,<0.116",
  "uvicorn[standard]>=0.30,<0.31",
  "pydantic>=2.8,<2.9",
  "asyncpg>=0.29,<0.30",
  "psycopg[binary]>=3.2,<3.3",
  "redis>=5.0,<5.1",
  "nats-py>=2.7,<2.8",
  "httpx>=0.27,<0.28",
  "sentence-transformers>=3.0,<4.0",
  "pgvector>=0.2,<0.3",
  "python-dotenv>=1.0,<2.0",
  "opentelemetry-sdk>=1.26,<1.27",
  "opentelemetry-instrumentation-fastapi>=0.47b0,<0.48",
  "opentelemetry-exporter-otlp>=1.26,<1.27",
  "tenacity>=9.0,<10.0",
  "python-json-logger>=2.0,<3.0",
  "python-jose[cryptography]>=3.3.0,<3.4.0",
  "jwcrypto>=1.5.6,<1.6.0",
  "aioboto3>=10.0,<11.0",
  "kubernetes-asyncio>=20.0,<21.0" 
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.7,<0.6.0",
  "mypy>=1.11,<1.12",
  "pytest>=8.3,<8.4",
  "pytest-asyncio>=0.23,<0.24",
  "pytest-cov>=5.0,<6.0",
  "httpx>=0.27,<0.28",
  "types-redis>=4.6.0,<5.0",
  "pip-audit>=2.7,<3.0"
]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu121"]  # Use PyTorch index for torch to avoid NVIDIA conflicts
index-strategy = "unsafe-best-match"  # Allow latest versions from all indexes (e.g., requests from PyPI)

[tool.ruff]
line-length = 100
target-version = "py311"
lint.ignore = ["E501", "F841"]
lint.select = ["E", "F", "W", "I", "N", "D", "UP", "S", "B", "A", "C4", "T20", "RUF"]
format = {quote-style = "single"}

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=xml --cov-report=term-missing"
